+++ node/src/components/fetcher/fetcher_impls/sync_leap_fetcher.rs
@@ -49,9 +49,9 @@ impl ItemFetcher<SyncLeap> for Fetcher<SyncLeap> {
                         .put_block_header_to_storage(Box::new(header.clone()))
                         .await;
                 }
-                for signed_header in item.signed_block_headers {
+                for block_header in item.block_headers_with_signatures {
                     effect_builder
-                        .put_signatures_to_storage(signed_header.block_signatures().clone())
+                        .put_signatures_to_storage(block_header.block_signatures().clone())
                         .await;
                 }
             }